cloud-function-analysis-simple/requirements.txt
google-cloud-storage==2.10.0
google-cloud-speech==2.21.0
google-cloud-aiplatform==1.38.0
functions-framework==3.4.0
requests==2.31.0
